ruse: Define a link action in the intermediate representation
A link action can be used to insert a runtime address into the
machine code for a Ruse function. For example, this might be
the address of another function or the address of a string
constant. The link structure is quite simple for now, but is
likely to become more sophisticated later.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>