ruse: Make full address link actions explicit
Previously, all link actions inserted the full address. This
commit doesn't change that behaviour, but makes it explicit
by introducing a link type (with only one initial value), which
is then used in a new Perform method to write the address.

This means that we can now add other link types and only have
to managed the behaviour in the Link type.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>