ruse: Encode link actions into instruction arguments
Previously, any link actions were passed around the side, stored
in assembler context and a temporary field in the instruction data.
This worked fine, but it was a little clumsy, and it would be
tedious for the compiler to reproduce when compiling Ruse down to
assembly.

Now, we store the link data in the argument, instead of the
placeholder value. This means looking through each of the arguments
where previously we would look at the aforementioned sidechannels,
plus updating the encoding code to look out for links and emit the
placeholder value instead. However, this is quite straightforward
and reduces the size of the instruction data.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>