ruse: Simplify instruction candidate handling to only use ssafir.Op
Previously, the generated mapping of mnemonic to instruction forms
would store the ssafir.Op and x86.Instruction for each option, despite
there being a one-to-one mapping between the two. This commit simplifies
that somewhat by instead storing just the ssafir.Op. This can then be
used to derive the instruction where needed.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>