ruse: Emit stack index registers in the x86 assembler
When we assemble instructions that act on x87 floating
point stack indices, we previously emitted just the
stack index as a uint8. This worked fine, but it was
less clear than other stored values. Instead, we now
store the register value. Although this is a bigger type,
it should not mean storing more memory, as it's stored
into an interface anyway. If anything, this may use
less memory, if the previous uint8 value was not stored
in place of a pointer.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>