ruse: Added a new implementation of x86 instruction generation
This adds an (as yet unused) alternative implementation for the
combination of //tools/ruse/internal/x86/(gen-x86 x86csv x86spec).
Subsequent commits will integrate the new code, replacing the old.

The current process evolved gradually in separate components.
x86spec parses the Intel x86 manual PDF, in code that has evolved
organically and developed technical debt. This outputs information
on x86 instructions in a CSV-based format. In the process, some
information is lost.

Next, gen-x86 parses the CSV, using x86csv, derives more information,
and adds more instructions. Then, the resulting instruction data is
written out in a JSON format and as Go data structures.

This approach works, but it's clumsy and inefficient, as it has to
rebuild in gen-x86 information that was known in x86spec but then
dropped.

This takes a more straightforward approach, combining the two tasks
into a single tool. The new code also tracks errors encountered in
the manual.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>