ruse: Added incomplete support for the A64 instruction set
I started having a go at building data on the A64 instruction
set, similar to the previous work on x86. Compared to x86, A64
has the advantage that Arm provides a structured representation
of the instruction set in XML format, which is much easier to
parse and test than the PDF parsing we have to do with x86.

However, the A64 instruction set is very different from x86
in ways I didn't expect. For example, operands and their
possible values often vary on a per-instruction basis, which
is very hard to parse programmatically.

A good example of this is ADD (extended register). The <extend>
operand has different values for LSL in 32- and 64-bit mode.
The <extend> operand is required, except if Rd or Rn is (W)SP.
This is all incredibly difficult to parse correctly, even with
structured data in XML.

As a result, I'm parking this and moving on to other work. I
may come back to it later, but it's low priority.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>