ruse: Fixed an assembler bug for memory offset segments
This fixes a bug where if an instruction included a memory
offset (such as MOV AL, moffs8) and the memory offset has
an explicit segment register, this would be parsed correctly
but the corresponding prefix would be omitted from the
machine code. This was caused by a check for memory
operands that omitted memory offset operands.

The fix is straightforward. At the same time, we now stop
iterating once we hit a nil operand, which is a minor
performance improvement.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>