ruse: Remove the requirement to include an architecture in asm-func
Previously, we required all assembly functions to specify their
architecture, using the '(arch) annotation. This was done on the
basis that assembly is unlikely to be valid in more than one
architecture.

However, this would be thoroughly tedious for any project that
needed to use lots of assembly and was only written for one
architecture. As a result, we now just assume the current
architecture if one is not specified.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>