ruse: Exclude parameter names from function type signature names
This is in the context of the type system, specifically the
*types.Signature type, which represents a function signature.
As with any type, this includes a name, which is a textual
description of the type.

Previously, we recorded the name of each parameter as part of
the function signature. For example, the Ruse function

	(func (binary-add (x int32) (y int32) int32)
		(+ x y))

Would have a type signature with the name

	(func (x int32) (y int32) int32)

This is helpful for developers, as it gives more context to
each function signature when debugging. However, it is slightly
misleading, as it gives the impression that the parameter names
are part of the type signature. This suggests that the above
signature type would be incompatible with

	(func (foo int32) (bar int32) int32)

which is incorrect.

As a result, this commit includes only the type information in
the name. For both of the above, the signature name would now
be

	(func (int32) (int32) int32)

This should make it easier to determine compatible types.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>