ruse: Added support for an object file format: rpkg
This introduces an object file format for Ruse, called rpkg. An
rpkg file is produced by compiling a single Ruse package. One or
more rpkg files can later be linked together into an executable
binary. This is the equivalent of a .o file for C or a .a file
for Go.

The rpkg file format is reasonably well-documented and the tests
include various samples. There is a 60-byte header, followed by
the following contiguous sections:

- Imports
- Types
- Symbols
- Strings
- Linkages
- Code

Finally, there is a SHA-256 checksum of the rest of the rpkg
file. Each section (and the header and checksum) is 32-bit
aligned by internal padding.

The main distinguishing feature of the rpkg format (and the
reason for its creation) is that it stores symbol names much
more efficiently for Ruse than the traditional object file
formats. This is because traditional formats were designed
for C-like symbols, which have a single namespace. In a
multi-namespace language like Ruse, this would produce
wildly inefficient symbol tables, where each symbol's package
name would be repeated in every symbol.

The rpkg format solves this by storing package names and
local names separately, performing de-duplication in the
strings section. This means that each package name is stored
in one place and referenced using its index in the strings
section.

As with Go archives, rpkg files include all of the types
and constants necessary to build the package, including
those imported from other packages. This means that when
compiling a package, we need only read from the rpkg files
of the direct dependencies. Any transitive dependencies
that could affect the build will have the relevant data
included in the direct depentancies' rpkg files. However,
the complete set must be included when linking, as any
transitive dependencies' functions' implementations will
only be in their rpkg files.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>