ruse: Added a link command to link rpkg files into a binary
This completes the work in the previous commit by adding the
ability to link together one or more rpkg files into an
executable binary. This involves most of the work that was
previously done by the compiler, but acting on information
from decoded rpkg files, rather than compiling source code.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>