ruse: Expose type checking as a package function, not method
Inspired by "go/types", we exposed the main type checking function
as a method on the Config type. However, we never used the Config
functionality, and it was inconsistent with the wider set of Ruse
packages.

This commit fixes that by changing the type checking function into
a package-level function.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>