ruse: Record exports in rpkg files
This includes the set of symbols exported from a package
in the rpkg file. This can then be used to recreate the
exposed package scope for a package. This is necessary
for us to import a package.

This also enriches the symbol handling in rpkg files,
which is used in the exports handling.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>