ruse: Record the compiled package versions into binaries
This is a simple but nice addition. For each package compiled into
a Ruse binary, we now record the package path and the cryptographic
checksum of its rpkg into the binary. These are stored in an 'rpkgs'
section, with a 16-bit length prefixed package path and fixed-size
(SHA-256) checksum. This can be extracted with tooling, or at run
time to give information about the provenance of the binary.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>