ruse: Emit ELF section headers as well as program headers
This extends the metadata included in the ELF binaries we produce.
Previously, we would only include program headers, which are the
headers used at runtime to instruct the loader which data needs to
be copied into memory.

This commit adds section headers, which give extra metadata prior
to runtime for analysing the binary. In particular, it gives named
sections, which can be useful. This also requires including the
NULL section header (which is required for some reason) and data
with the section names (and a corresponding section header).

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>