ruse: Added initial support for ABIs
Previously, we had a weak form of calling convention support for
assembly functions. However, this, didn't extend to Ruse functions
and had to be provided explicitly for each function.

This commit starts by adding explicit support for ABIs in the
sys package, including default ABIs for both x86 and x86-64 (the
initial defaults are System-V). This gives a way to express the
distribution of registers to parameters, results, scratch registers
(caller-preserved) and unused registers (callee-preserved). It
also allows custom ABIs to be specified.

Next, we add the functionality to derive a function's calling
convention from its ABI (or the default ABI if needs be). We have
some basic tests for this too.

Next, we add ABI support to the types package. Custom ABI specs can
now be defined using the (abi) special form. These can be stored in
constants or provided inline. An ABI can now be provided in an '(abi)
annotation to a function (either Ruse or assembly). The resulting
ABI is included in the types.Function object.

Finally, the compiler now uses the ABI to derive each function's
calling convention. We don't use the calling convention yet, but
that will follow shortly.

The next step will be to remove the old calling convention code
for assembly functions, converging on the single ABI mechanism.
This will also unify the function signature syntax.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>