ruse: Unify function signature syntax
This commit converges both assembly and Ruse functions to the
same syntax for the function signature. Now that we have proper
support for ABIs, we can drop the old support for specifying the
calling convention of an assembly function. Having done so, we
can now use the same syntax for function signatures in assembly
functions that we use in Ruse functions.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>