ruse: Require explicit ABIs when calling functions from assembly
With our new ABI support, we have the default ABI for each architecture.
However, this default ABI is not guaranteed to be stable. As a result,
we do not want to allow assembly code (which is not fully validated by
the compiler) to call functions that use the default ABI.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>