ruse: Exclude parameters and results from unused registers in ABIs
Previously, we erroneously used only the set of scratch registers
to derive/invalidate the set of unused registers. This commit
fixes that by considering the parameter and result registers too.
This is because we can't preserve registers values that are
populated by the caller with parameters or that we populate with
results.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>