ruse: Add a primitive register allocator
This currently only supports a few basic operations, but it can
allocate registers in an architecture-abstract way and includes
tests for the supported basic operations.

The implementation works by replacing ssafir.Value instances
for abstract instructions with concrete versions that contain
register allocation information.

The biggest issue at the moment is that it compresses the
blocks in a function into a single, linear block. Once we start
to support branching code, that will need to be undone to preserve
the block structure.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>