ruse: Add data on registers with different sized segments
Many x86 registers have multiple versions where different
sized segments of the register can be referenced. For
example, the `a` register can be referenced with:

- `al` (bottom 8 bits)
- `ax` (bottom 16 bits)
- `eax` (bottom 32 bits)
- `rax` (bottom 64 bits)

The new `RegisterSegments` variable contains these groups
of register segments and `Register.ToSize` allows one register
to be replaced with a different sized segment of the same
underlying register.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>