ruse: Add the option to include a symbol table
This adds functionality to binary/elf to encode the provided symbols
in a standard symbol table. Unfortunately, this is fairly inefficient
on space in the binary, as the symbol table and the table of symbol
names are both page-aligned. This is due to the existing logic on keeping
sections in separate pages to simplify alignment. However, if space is
important, we can just omit the symbol table.

Note that this also includes fixing the symbol table addresses and
offsets in elf.Encode, rather than in cmd/link. This is necessary so
that we can encode the symbol addresses correctly in the symbol table.

We also add a test case that is otherwise the same, to check that the
code is correct whether we include the symbol table or not.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>