ruse: Add the ability to store named ABIs in rpkg files
Ruse has the ability to declare a named ABI by assigning an ABI
object to a constant. However, we didn't previously have a way
to store such a constant in an rpkg file. This commit adds that
ability, along with an example test.

We store the ABI in the constant's type, along with a zero
value. The other machinery is already in place.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>