ruse/types: Check all annotations in a consistent way
Previously, we checked some annotations in places where they were
supported. This wasn't applied consistently or completely. This
commit adds new functionality that checks the entire syntax tree
for annotations in a consistent manner. We also add some tests.

Note that we generally ban the use of annotations that we have
not yet implemented, to ensure forward compatibility.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>