ruse/binary: Use an entry point symbol, not a base address
Previously, we would supply an explicit base address and bank on
the fact that that would currently be the same address as the
main function in the code section. If we changed the order of the
sections or didn't put the main function at the beginning of the
code section, this would break painfully.

In this commit, we change that to instead referencing the symbol
for the entry point, which will typically be the main function.
Now that the symbols and sections have their offsets and addresses
correct before we start writing the ELF header, we can simply pull
the entry point address from the entry symbol. This also means that
the order of the sections (and the location of the entry point
within those sections) is now decoupled.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>