ruse/cmd/link: Move section order into data
Previously, we made assumptions about the relative order of the
different sections in the binary, particularly when linking each
symbol to its parent section.

This commit puts the section definitions up front, building a mapping
from each to its index. This is then used to join up the mappings of
symbols to sections. Finally, each section is populated with data and
added to the binary. This means that changing the order is just a case
of changing the sections slice order.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>