ruse/pkg/link: Refactor symbol handling
Previously, we would declare the set of section, process the main
package and its symbols, then do the same for the dependencies and
standard library.

Now, we collect all of the packages, writing them into the relevant
section data, then declare the sections, and finally process the
package symbols. The effect is the same, but with less duplicate
code and without needing to declare the sections until they're
needed.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>