ruse/cmd/link: Combine the final section handling in the linker
Previously, we had one last inconsistency in our section handling,
which is that we had the ability to provide a custom section symbol,
but would always use the default section for section indices in the
symbol table.

This unifies the handling into a single place, which is also more
concise.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>