ruse/types: Allow symbols to specify their section
Previously, symbols always exist in their default section. This
adds a new section annotation, which can be used with a reference
to a named section to state that the symbol should be linked into
that section. For now, this just references the section by name.
Support in the linker will follow.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>