ruse/compiler: Add a package annotation to specify program sections
This syntactic support and type checking for a new annotation on
the package statement that provides the ordered set of program
sections that the linker should use. This allows the package to
change the order of the sections and to omit sections, removing
them from the linked binary.

Note that this is purely syntactic and types support. It does not
yet have any effect on the linker.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>