ruse/cmd/link: Use any specified set of sections while linking
This completes the support for custom sections. If package main
specifies a set of sections, we only include those sections
(which may contain one section more than once) in the order
specified, omitting any other sections. This can be used to
drop unwanted sections and to control the order and addresses
of each section.

Note that this will not prevent critical sections from being
removed. For example, if sections.Code is omitted and any
necessary code is not in another section, the binary will
likely fail at runtime.

Note that this can safely be used to omit the rpkgs provenance
data by omitting the rpkgs section.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>