ruse/compiler: Add basic support for type casts
This adds primitive support for type casts. If a function call
is made on a type name, with one argument, which is assignable
to that type, then we echo the value but change the type. It's
quite possible that this will fail awkwardly when the type
system becomes more complex, but it works for now.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>