parser: Fixed handling of unattached annotations
We require that annotations must be attached to either a list or
another annotation. That is, there must not be a blank line between
the annotations on a list.

Previously, our handling of this was too primitive and only worked
for annotations that are one or two lines long. This corrects that
by tracking the previous annotation and comparing against its closing
parenthesis.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>