sys: Added data on register families
In many architectures, there are child registers that form part
of one or more parent registers. In such cases, a write to the
child register will also affect every parent register. In some
cases, parent registers can themselves be child registers of an
ancestor register.

For example, in x86, the AL register is a child of AX, which is
a child of EAX.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>