ssafir: Correctly handle checks on signed address offsets
When we link a symbol address offset into a symbol, we check that
the address offset fits in the space available. Previously, we
assumed that offsets were always positive. If they were negative,
then the resulting unsigned value would always be truncated in
16-bit and 32-bit stores.

This commit fixes that by checking whether the signed value would
be truncated as well.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>