ruse: Allow 16-bit relative addresses in 16-bit mode
Previously, we would always require 32-bit relative addresses for
instructions like jmp and call, in case the destination was far
away. However, in 16-bit mode we can be confident that the destination
will be near and than therefore allow 16-bit relative addresses.

This also prevents us from producing 32-bit jumps/calls in 16-bit
mode, which may be allowed by the processor, but may not.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>