ruse: Make + and - special identifiers, which must be detatched
Previously, when we encountered a plus or minus, we would check whether
it was followed by a digit (in which case return a number), or otherwise
return an identifier. This functioned correctly, but would allow quite
open-ended identifier forms that might be confusing, like `-foo` and
`+seven`.

This commit specialises the form, requiring that the plus/minus is not
followed by any more identifier runes.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>