bootimage: Refactored the bootimage to focus on just the bootloader
This is a refactor of the original Firefly bootimage tool. The main
change is that we only process the bootloader binary now; we don't
yet take a kernel binary or user package.

The other main change is that we're slightly more tightly coupled
to our bootloader (//bootloader) and we insert the MBR magic on its
behalf.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>