ruse/types: Added a set of tests for individual identifiers
The TestCheck tests are quiet clumsy, as we have to reproduce the
full type information, along with positions. This adds another set
of tests (only a few for now), which just check the type (and
optionally a constant value) for a single identifier. This allows
us more flexibility to add context, without then having to produce
vast amounts of data.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>