ruse/types: Added support for an alignment annotation
This adds a new alignment annotation that can be used on constant
and function declarations (including assembly functions). This
will (in future) allow the const/func to be constrained to the
given alignment in memory. For now, this just makes the type-checker
aware of the annotation. It does nothing yet.

Example:

	'(align 32)      ; Align to a 32-byte boundary in memory.
	(let foo "test")

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>