bootloader: Embed the MBR magic in the bootloader directly
Previously, we inserted the MBR magic into the bootloader binary
using the bootimage tool. That worked fine, but felt a little
unnecessary. Now that we have support for constant alignment and
store integer constants into the binary, we can just declare the
MBR magic as a constant with the right alignment. This simplifies
the bootimage tool.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>