ruse/types: Allow numerical constants to be inserted in assembly
This adds the ability to insert Ruse integer constant expressions
into assembly code. The expression must result in a constant value
that is assignable to the stated type, but there is otherwise no
restriction on the expression. This will be a useful tool in
ensuring that Ruse assembly code remains consistent with other
Ruse code, by allowing the use of shared constants.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>