bootloader: Load the stage 2 end address from a named symbol
When we copy the rest of the bootloader into memory, we need to
know where the bootloader stops. Previously, we did this by writing
the end address of the bootloader to a fixed offset into the binary
and read it from that same address. This worked well enough, but
felt clumsy and brittle.

This commit changes to using a named symbol to determine the address
in memory where the end address would be written.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>