bootimage: Preserve spacing between bootloader sections
When we write the bootloader to the disk image, we should preserve
any gaps between sections, so that the resulting addresses/offsets
will still match. The result is that the image is identical to the
bootloader in memory, except it starts at the first section (which
is at 0x7c00).

The previous behaviour was incorrect, but went undetected as we
only had two sections, which were adjacent in memory already.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>