bootloader: Specify the package base address, rather than sections
This commit switches from one valid approach to another. Previously,
we specified the sections that would be compiled and their order,
which would safely result in the memory layout we want. This specifies
the base address, which will result in the same thing. However, this
new approach will be more flexible if we use more sections.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>