bootloader: Use standard sections after stage 1
Previously, we used an explicit section for stage 2 of the bootloader.
This worked fine but was quite verbose (and unnecessarily put strings
and other data in an executable section).

This switches to using the standard default sections for the rest of
the bootloader, which is a bit simpler. It does use slightly more
memory (as the linker aligns sections to 4 kiB boundaries), but that
feels like an acceptable trade-off.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>