ruse/compiler: Moved the test vectors into a separate target
The x86 test vectors are expensive to run. The current approach
of using a command-line flag to gate running them is tedious to
use. This moves it into a separate Bazel test target. It means
the Go tooling would run them by default, but we don't use it
any more.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>