ruse/compiler: Allow short jumps based on the RCX register
As a general rule, we only select 32-bit jump instructions,
as they tend to have short encodings (so get picked) but may
then not have enough space to encode the jump distance,
leaving us stuck.

This adds an exception for the jrcxz, jecxz, and jcxz
instructions, as they only have 8-bit jump forms.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>