ruse/compiler: Preserve registers overwritten by function args
When we perform a function call, we need to move its arguments
into the relevant locations specified by the function's ABI.
This may involve overwriting values already in those locations.

This commit ensures that when we do so, we preserve the previous
values in those locations, so they are not lost.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>