ruse/compiler: Add the desired assembly in lowering tests
It can be tricky to understand the lowered instruction data that
we produce. To make the tests easier to read and understand, we
now use golang.org/x/arch/x86/x86asm to print the disassembly of
the assembled instructions and compare against test data. This
also ensures that our finally assmebled code is correct.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>