ruse/compiler: Group together multi-stage arithmetic operations
When we do a multi-stage arithmetic operation, like (+ 1 2 3), we
split it into separate operations like (let tmp (+ 1 2)) and (+ tmp 3).
Previously, we would use the standard NewValue function for each
sub-operation. This means that each sub-operation will have a unique
ssafir.ID.

In this commit, we make a slight tweak to use the same ssafir.ID for
every sub-operation. This can't do any harm, as there's no way to
refer to a sub-operation from elsewhere in the code and we already
have cases where a single operation results in multiple values with
the same ID.

We then make a second change so that when the first argument to an
arithmetic operation has the same ID, we reuse its memory location,
as we now know that won't cause any problems. This means we can use
fewer memory locations and skip some otherwise unnecessary moves.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>