ruse/compiler: Use × for multiply and ÷ for divide, with aliases
This changes the name of the multiply function to × (but retaining
an alias *) and the divide function to ÷ (with an alias /). This is
compatible with previous code, for those who prefer sigils on a
typical keyboard. Personally, I find the new symbols easier to
understand at a glance with the S-expression style.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>