ruse/compiler: Refactor allocation/lowering, using op metadata
This is a reasonably large refactoring job, which should have little
functional impact (except perhaps a slight performance improvement
by replacing a loop with a lookup).

Broadly, this uses groups to collapse large sets of operations and
size information to simplify instruction selection. The result is
less repetitive and should be easier to maintain over time. It
should also be much less prone to subtle errors.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>