ruse/compiler: Take a better approach to preventing keywords
Rather than doing this up in the compiler, we now check in the
type checker, as it can simply look up the new name in the set
of special forms. This is easier to maintain over time and less
likely to have subtle errors.

The fact that we've had to fix some tests as part of this commit
demonstrates that the previous approach wasn't working.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>