ruse/compiler: Add the ability to load relative addresses with LEA
Currently, if we want to store an address or string pointer into
a register, we use MOV, storing the address as a 64-bit immediate.
This adds a better option, which is to use LEA with a RIP-relative
address. This uses a relative address (so works better with ASLR)
and encodes to a smaller instruction (7 bytes, rather than 10).

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>