ruse/compiler: Fixed scratch register avoidance when calling funcs
Previously, we would check whether we need to move any data out of
registers before calling a function, based on the registers that
function trashes. However, we would then prepare the parameters for
the function, involving subsequent moves. These latter moves could
accidentally move data into registers that the function would then
trash.

This commit fixes that by passing on the set set of registers to
avoid when preparing parameters.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>