ruse/link: Added a basic tree shaker to remove unused symbols
This is a fairly straightforward implementation that builds up the
dependency graph of symbol references, then keeps only those that
have a path to the main function. We also keep ABIs and sections,
as tracking them is tricky and they don't bloat the binary if unused.
We also keep constants with a fixed alignment, as they may exist
only to belong in memory.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>