ruse/link: Don't bubble up transitive dependencies in tree shake
Previously, when we store the dependency graph, we would bubble
it upwards, including (some of) each link's transitive dependencies.
However, this was incomplete as we ensured the right order of processing
at the package level, but not the function level. This meant that
if package 'foo' depends on package 'bar', the latter would be
processed first, but no such ordering was applied within each
package.

Since this bubbling wasn't complete, we didn't rely on it and
traversed the full graph. Since we traversed the full graph, there
is no benefit in doing any bubbling, as it simply results in
duplicated effort and storage.

As a result, this commit drops that.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>