ruse/compiler: Add ENDBR instructions to function preludes
This commit changes the compiler and assembler to add an extra
instruction at the beginning of each x86-64 function. In 64-bit
functions, we add ENDBR64, while in 32-bit functions, we add
ENDBR32. 16-bit functions remain unchanged.

In future, this should enable us to support CET Indirect Branch
Tracking. We currently only support calls to the beginning of
each function, so that's where we place the landing instructions.
Jumps within a function use fixed-size jumps and thus are not
affected by IBT. If we want to jump or call within a function in
assembly, we can insert the ENDBR instruction manually.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>