ruse/format: Rewrote the formatter
This is a near-complete rewrite of the formatter. The result is
much better tested and more consistent. The main stylistic changes
are that it gives much greater deference to the input source for
where to place newlines. We still compress runs of newlines and
force function signatures to be on the same line as the keyword
(and to include each assembly instruction on its own line), but
otherwise we accept the input's choice.

The other main change is that we now handle line comments correctly.
Previously, they were dropped entirely. This has been tested on the
whole bootloader, which is a good starting point.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>