ruse/compiler: Add support for multiple returns, multi-value let
This makes two main changes. Firstly, it follows the previous commit
in adding full support for functions returning multiple values. For
now this can actually only be performed in assembly functions, but
the compiler will handle it correctly.

The second change is to allow let statements in a function to store
multiple returns from a single function to an equal number of new
variables. Note that we can't assign separate values to separate
names, as the syntax would be ambiguous.

Example demonstration:

	(package main)

	(func (main)
	        (let x2 x4 (x2-and-x4 3))  ; Store 3x2 in x2 and 3x4 in x4.
	        (exit (+ x2 x4)))          ; 3x2 + 3x4 = 18

	; The syscall ABI.
	;
	(let ABI (abi
	        (params rdi rsi rdx r10 r8 r9)
	        (result rax rdx)))

	'(abi ABI)
	(asm-func (exit (code int))
	        (mov eax 60)  ; sys_exit
	        (syscall))    ; exit(code)

	'(abi ABI)
	(asm-func (x2-and-x4 (n int) int int)
	        (mov rax rdi)
	        (mov rdx rdi)
	        (sal rax 1)
	        (sal rdx 2)
	        (ret))

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>