ruse/compiler: Added support for return statements
This defines return statements, including multiple returns. The
behaviour is as you'd expect; a return statement works like any
other and must take the right number (and type) arguments as the
result of the innermost parent function.

Using an explicit return statement (or ending with a function
call of the same result type) is the only way to return multiple
values in Ruse code, so it's an important change.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>