ruse/compiler: Added an optimisation for shifting by a constant
Previously, if we did a shift by a constant value, we would store
the constant to RCX and then use the usual SAL/SAR_RmrX_CL. This
commit adds an optimisation where instead we either use the special
SAL/SAR_RmrX_1 form (if the shift is the constant 1) or the
SAL/SAR_RmrX_Imm8u form with the shift included as an immediate.

This isn't hugely important, but it's a good first example of
specialising an instruction.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>