ruse/compiler: Added basic support for if statements
This adds initial support for if statements. The functionality in
the formatter, type checker, compiler, and lowerer should be solid.
However, support in the allocator is more basic. In particular, we
currently handle the if block (and optional else block) as continuing
from the previous block. This means that the former affects the
latter for allocations and some values may not be dropped correctly.
Fixing this will be tricky, but it's already a pretty big commit.

Similarly, we do (in theory) support using an if statement as a
value, provided the if and else blocks result in the same type.
However, getting that to actually work in the allocator will be
hard.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>