ruse/compiler: Use allocator snapshots to fix if block handling
Our initial support for if statements had a simple limitation.
The allocator treated the if and else blocks as normal blocks
that are always both executed. As a result, the else block used
any changes that happened in the if block and the final block
used any changes in the else block.

This commit fixes that by introducing the concept of allocator
snapshots and using them to restore state between blocks. It
also includes tests for both producing and reverting snapshots.

A snapshot includes a representation of the current allocator
state. In particular, it stores the data held in a.allocated
and a.locations. It does so in a deterministic manner, so that
the behaviour is repeatable.

Reverting a snapshot identifies the moves necessary to restore
the previous state, while also checking that the allocator does
not have any additional values still live.

We take a snapshot before the if block and revert it at the end
of the if (and else, if there is one) block.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>